requirements.txt
pywebview==4.4.1
py2app==0.28.6
